+++ CHANGELOG.rst
@@ -5,6 +5,16 @@ Changelog for the **SODAR Core** Django app package. Loosely follows the
 `Keep a Changelog <http://keepachangelog.com/en/1.0.0/>`_ guidelines.
 
 
+Unreleased
+==========
+
+Changed
+-------
+
+- **General**
+    - Upgrade minimum Django version to v4.2.25 (#1774)
+
+
 v1.2.3 (2025-09-26)
 ===================